@@ -16,7 +16,7 @@ namespace EngineBay.Persistence
 
             var databaseProvider = (DatabaseProviderTypes)Enum.Parse(typeof(DatabaseProviderTypes), databaseProviderEnvironmentVariable);
 
-            if (!Enum.IsDefined(typeof(DatabaseProviderTypes), databaseProvider) | databaseProvider.ToString().Contains(',', StringComparison.InvariantCulture))
+            if (!Enum.IsDefined(typeof(DatabaseProviderTypes), databaseProvider) || databaseProvider.ToString().Contains(',', StringComparison.InvariantCulture))
             {
                 Console.WriteLine($"Warning: '{databaseProviderEnvironmentVariable}' is not a valid {EnvironmentVariableConstants.DATABASEPROVIDER} configuration option. Valid options are: ");
                 foreach (string name in Enum.GetNames(typeof(DatabaseProviderTypes)))
@@ -47,7 +47,7 @@ namespace EngineBay.Persistence
                 return false;
             }
 
-            if (databaseResetString.ToUpperInvariant() == "TRUE")
+            if (string.Equals(databaseResetString, "TRUE", StringComparison.OrdinalIgnoreCase))
             {
                 Console.WriteLine($"Warning: {EnvironmentVariableConstants.DATABASERESET} was set to 'true', this will RESET the database to default. I hope you know what you're doing...");
                 return true;
@@ -65,7 +65,7 @@ namespace EngineBay.Persistence
                 return false;
             }
 
-            if (databaseReseedString.ToUpperInvariant() == "TRUE")
+            if (string.Equals(databaseReseedString, "TRUE", StringComparison.OrdinalIgnoreCase))
             {
                 Console.WriteLine($"Warning: {EnvironmentVariableConstants.DATABASERESEED} was set to 'true', this will attempt to insert default data into the database. I hope you know what you're doing...");
                 return true;
@@ -83,7 +83,7 @@ namespace EngineBay.Persistence
                 return false;
             }
 
-            if (exitAfterMigrationsString.ToUpperInvariant() == "TRUE")
+            if (string.Equals(exitAfterMigrationsString, "TRUE", StringComparison.OrdinalIgnoreCase))
             {
                 Console.WriteLine($"Warning: {EnvironmentVariableConstants.DATABASEEXITAFTERMIGRATIONS} was set to 'true', this will exit the process once database migrations have been applied.");
                 return true;
@@ -101,36 +101,13 @@ namespace EngineBay.Persistence
                 return false;
             }
 
-            if (exitAfterSeedingString.ToUpperInvariant() == "TRUE")
+            if (string.Equals(exitAfterSeedingString, "TRUE", StringComparison.OrdinalIgnoreCase))
             {
                 Console.WriteLine($"Warning: {EnvironmentVariableConstants.DATABASEEXITAFTERSEEDING} was set to 'true', this will exit the process once database seed data has been inserted.");
                 return true;
             }
 
             return false;
-        }
-
-        public static bool IsAuditingEnabled()
-        {
-            var auditingEnabledEnvironmentVariable = Environment.GetEnvironmentVariable(EnvironmentVariableConstants.DATABASEAUDITINGENABLED);
-
-            if (string.IsNullOrEmpty(auditingEnabledEnvironmentVariable))
-            {
-                return true;
-            }
-
-            bool auditingEnabled;
-            if (bool.TryParse(auditingEnabledEnvironmentVariable, out auditingEnabled))
-            {
-                if (!auditingEnabled)
-                {
-                    Console.WriteLine($"Warning: Auditing has been disabled by {EnvironmentVariableConstants.DATABASEAUDITINGENABLED} configuration.");
-                }
-
-                return auditingEnabled;
-            }
-
-            throw new ArgumentException($"Invalid {EnvironmentVariableConstants.DATABASEAUDITINGENABLED} configuration.");
         }
 
         public static string GetDatabaseConnectionString()
