@@ -2,10 +2,8 @@ namespace EngineBay.Persistence
 {
     public interface IModuleWriteDbContext : IModuleQueryDbContext
     {
-        int SaveChanges(ApplicationUser user);
+        int SaveChanges();
 
         new Task<int> SaveChangesAsync(CancellationToken cancellationToken = default);
-
-        Task<int> SaveChangesAsync(ApplicationUser user, CancellationToken cancellationToken = default);
     }
 }
